Add platform specific behavior to layout
This commit introduces a platform check to determine if the WindowTitlebar should be displayed. Now, the WindowTitlebar, including all its control classes and attributes, is only displayed if the platform is not android. Also, an input element for search and its accompanying shortkey text has been added to the WindowTitlebar.