Add PocketBase support and types
This commit introduces the PocketBase library into the project, along with creating specific types and classes for handling its service (pocketbase-types.ts). The changes made here enable the usage of PocketBase in the system. This is reflected in the update to package.json where the PocketBase version has been specified.